Refactored intcode program from array to structure including pc/relative base

+++ 2019/sol/include/sol.h
@@ -17,7 +17,7 @@ int solve_2019_3_2(char const *input_filename);
 int solve_2019_4_1(char const *input_filename);
 int solve_2019_4_2(char const *input_filename);
 
-long solve_2019_5(char const *input_filename, int input);
+long solve_2019_5(char const *input_filename, long input);
 
 int solve_2019_6_1(char const *input_filename);
 int solve_2019_6_2(char const *input_filename);
